monitoring/requirements.txt
evidently
azure-storage-blob
pandas
